GUI Runner: Don't x/0 when calculating fps
It seems that runner->totalDelta has the potentially to overflow back to
0, thus resulting in x/0. This change makes sure that this is never the
case.

This is required for the TI Nspire platform, because for some reason,
trying to format an Infinity (what x/0 results) bugs out the formatter,
corrupting memory with mainly '0' (0x30), prepended with an "Infinity"
(and "NaN\0"..?).

At least I had some 'fun' playing Space Invaders trying to reproduce.