TINspire: Align sizes of binaries
On real hardware, the device crashes upon reading ext_icons_size because
it wasn't aligned. ext_font_size doesn't have this problem because the
image conveniently ends with the correct alignment, but in the case it
changes, we're secure.

This was a pain to debug, since you don't really see crashing
"between the functions" everyday, and when replacing the parameters and
seeing that putting ext_icons_size crashes, I wondered if ext_icons_size
wouldn't be aligned correctly and thus cause an interrupt (which just
restarts), since from what I've heard ARM normally wants aligned
addressed compared to, say, x86.

All of the testing till then was done on an emulator, firebird, which
has some inaccuracies, but generally works, and has gdb support.